Logging Documentation

The functions define here assist with debugging by logging some results to disk.

Detailed Documentation

Utils.sharedFaceLogging — Method.

Utils.sharedFaceLogging

This function writes files for the function that evalutes the shared face integrals. This function should only be called in debug mode.

If opts["writeqface"] is true, writes the q values at the face to a file q_sharedface_i_myrank.dat, where i is the peer process index (not MPI rank) and myrank is the MPI rank of the current process.

If opts["write_fluxface"] is true, it writes eqn.flux_sharedface to a file flux_sharedface_i_myrank.dat

Inputs: mesh sbp eqn: and AbstractSolutionData opts: options dictonary data: SharedFaceData object qL_arr: an array holding solution values at face nodes on the left side of the interface, numDofPerNode x numfacenodes x numsharedfaces on this partition boundary qR_arr: solution values at face nodes on the right side of the interface. same shape as qL_arr

also, the eqn.flux_shared is used to write the face flux. It is the same shape as qL_arr

Aliasing restrictions: qL_arr, qR_arr, and eqn.flux_sharedface must not alias.