Documentation of the PDESolver Physics Module Interface

Evaluating the Physics

Registration Functions

These function provide the basic API for registering and retrieving physics modules with the PDESolver module.

PDESolver.register_physics — Function.

This function registered a new physics module with the global list of all known physics modules. Every physics module should do this as part of module initialization. The name, the module, and the startup function must be unique (ie. they must not already exist in the list). This function throws and exception if they are not.

Inputs:

modname:  an ASCIIString name for this entry in the list.  It is used
          to retrieve the module and startup function in the 
          retrieve_physics function. Typically the name is capitalized.
mod:  the Module itself
startup_func: the function for running the physics.  It must have signature
              startup_func(fname::ASCIIString), where fname is the name of
              an input file

Outputs:

none

PDESolver.retrieve_physics — Function.

Retrieves the physics module and function registered using register_physics

Input:

modname: an ASCIIString containing the name of the module supplied to
         `register_physics`

Outputs:

mod: the physics Module
func: the function to evaluate the physics

See also registerIC and registerBC.