Starting a Simulation

This page describes functions located in the PDESolver module that tie together the physics modules and the Nonlinear solvers.

PDESolver.run_solver — Method.

This function provides a way to invoke any physics solver based on the specification of the physics in the input file. This requires loading the input file twice, once to figure out the physics, and a second time when the physics-specific startup function is called

The physics module must have already been registered using register_physics

Inputs:

input_file: an AbstractString specifying the path to the input file

Outputs:

mesh: the AbstractMesh object used during the solve
sbp: the SBP operator used by the solver
eqn: the AbstractSolutionData object during the solve.  At exit,
     eqn.q_vec should have the final solution in it
opts: the options dictionary

PDESolver.call_nlsolver — Function.

This function takes in the 4 principle object, fully initialized, and calls a nonlinear solver on them, according to the options in the dictionary. The evalResidual function is passed to the nonlinear solver

Inputs: mesh: a mesh object sbp: an SBP operator eqn: an equation object opts: options dictionary, used to determine which nonlinear solver to call pmesh: mesh used for calculating preconditioning jacobian in Newton's method, default to using mesh if not specified

Outputs: none

Aliasing restrictions: none (specificaly, mesh and pmesh can be the same object)

PDESolver.createMeshAndOperator — Method.

Create a SBP operator and a mesh. This is used by all physics modules to create the right type of operator and mesh based on the input options. It is type unstable, but that is ok.

If the options dictionary specifies a second SBP operator type, a second mesh and SBP operator will be created and stored in the mesh2 and sbp2

Inputs: opts: options dictonary dofpernode: number of degrees of freedom on each node

Outputs sbp : an AbstractSBP mesh : an AbstractMesh pmesh : an AbstractMesh, used for preconditioning, may be same object as mesh Tsol : DataType that should be used for eqn.q Tres : DataType that should be used for eqn.res Tmsh : DataType of mesh.dxidx and friends mesh_time : time in seconds for creation of mesh (Float64)

PDESolver.createMesh — Function.

This function creates the mesh object and, optionally, a second mesh used for preconditioning

Inputs: opts: the options dictionary sbp: an SBP operator sbpface: an SBP face operator topo: an ElementTopology describing the SBP reference element. Only needed for 3D DG, otherwise can be any value Tmsh: the DataType of the elements of the mesh arrays (dxidx, jac, etc.) dofpernode: number of degrees of freedom on every node suffix: suffix added to options dictionary keys that describe the SBP operator. See createSBPOperator

All arguments except opts are typically provided by createSBPOperator and getDataTypes

PDESolver.createSBPOperator — Function.

This function constructs the SBP operator and the associated SBP face operator, as specified by the options dictionary. It also determines the shape_type that PumiInterface uses to describe the SBP operator to Pumi.

Inputs: opts: the options dictionary Tsbp: the DataType specifying the Tsbp passed to the SBP operator constructor suffix: this suffix is added to all keys accessed in the options dictionary. Usually the suffix is either the empty string or an integer. This provides a convenient way for the input file to specify several different SBP operator and have this operator construct them. Default value is the empty string.

Outputs: sbp: the SBP operator sbpface: the SBP face operator shape_type: an integer passed to the mesh constructor to describe the operator topo: in the 3D DG case, an ElementTopology describing the SBP reference element, otherwise the integer 0.

PDESolver.getDataTypes — Method.

This function determines the datatypes of the elements of the arrays of the mesh quantities, sbp operator, solution variables and residual.

If the datatypes cannot be determined, an error is thrown.

Inputs: opts: the options dictionary

Outputs Tmsh Tsbp Tsol Tres

Physics Module Startup

TODO: does this belong here, or in the physics module section?

Each physics module is required to do some of the setup work needed to start a simulation. The functions above facilitate doing so. In particular, the physics module must

read the input dictionary

create an AbstractMesh and AbstractSBP

create an AbstractSolutionData

Load an initial condition

Calculate various quantities

Invoke a NonlinearSolver

Do postprocessing

Input Dictionary

Creating Mesh and Operator

Create an Equation Object

Load an initial condition

Various calculations

Invoke a NonlinearSolver

Do Postprocessing

The functions here should be used by all physics modules to assist in creating the AbstractMesh and AbstractSBP objects and calling a nonlinear solver.

TODO: document what the physics module needs to implement for startup