Advection Equation Steady Adjoint

PDESolver currently has the capability to compute the steady adjoint of a boundary functional. Recall the adjoint equation as

\[\frac{\partial \mathcal{L}}{\partial q} = \frac{\partial \mathcal{J}}{\partial q} + \psi^T \frac{\partial \mathcal{R}}{\partial q} = 0\]

where, $\mathcal{L}$ is the Lagrangian for functional $\mathcal{J}$ and $q$ is the solution variable. The adjoint can be computed by calling the function calcAdjoint, which has been described below.

AdvectionEquationMod.calcAdjoint — Method.

AdvectionEquationMod.calcAdjoint

Calculates the adjoint vector, ψ, for a single functional. Currently only DG meshes are supported. The function performs a direct solve using Julia's \ operator. For parallel meshes, a PETSc solve is done using ILU factorization. The user always call this function in order to compute the adjoint.

Inputs

mesh : Abstract mesh type

sbp : Summation-By-Parts operator

eqn : Advection equation object

opts : Options dictionary

functionalData : Object of type AbstractOptimizationData. This is the type associated with the adjoint of the functional being computed and holds all the necessary data.

adjoint_vec : Adjoint vector corresponding to the particular functional computed. If called in parallel, the vector should be distributed across eqn.comm, just like eqn.q_vec

functional_number : The functional for which the adjoint vector is being, default = 1

Outputs

None

AdvectionEquationMod.calcFunctionalDeriv — Method.

AdvectionEquationMod.calcFunctionalDeriv

Computes a 3D array of the derivative of a functional w.r.t eqn.q on all mesh nodes.

Inputs

mesh : Abstract mesh object

sbp : Summation-By-Parts operator

eqn : Advection equation object

opts : Options dictionary

functionalData : Object of subtype of AbstractOptimizationData. This is the type associated with the adjoint of the functional being computed and holds all the necessary data.

func_deriv_arr : 3D array that stors the derivative of functional w.r.t eqn.q. It has a structure [1, numnodes_per_element, numEl]

Outputs

None

AdvectionEquationMod.calcIntegrandDeriv — Method.

AdvectionEquationMod.calcIntegrandDeriv

Compute the derivative of the integrand at a point. It presently uses complex step to compute the derivative

Inputs

opts : Input dictionary

params : the ParamType for the equation

nx & ny : Normal vectors

q : Solution variable

functionalData : Functional object

Outputs

integrand_deriv : derivative of the functor w.r.t q

AdvectionEquationMod.calcResidualJacobian — Method.

###AdvectionEquationMod.calcResidualJacobian

The function calculates the residual for computing the adjoint vector. The function allows for jacobian to be computed depending on the jacobian type specified in the options dictionary jac_type.

Input

mesh : Abstract mesh object

sbp : Summation-By-parts operator

eqn : Euler equation object

opts : options dictionary

Output

jac : Jacobian matrix