Boundary Functional

Advection Boundary Functional

This page consists of all the functions necessary for computing a boundary functional along the geometric edges of a mesh for the advection equation.

AdvectionEquationMod.calcBndryFunctional — Method.

AdvectionEquationMod.calcBndryfunctional

This function calculates the functional on a geometric boundary of a the computational space. This is a mid level function that should not be called from outside the module. Depending on the functional being computed, it may be necessary to define another method for this function based on a different boundary functional type.

Arguments

mesh : Abstract mesh object

sbp : Summation-By-Parts operator

eqn : Advection equation object

opts : Options dictionary

functionalData : Object of the functional being computed

source

AdvectionEquationMod.calcBoundaryFunctionalIntegrand — Method.

Method for IntegralQData functional.

Note that q should be scaled by the length of the normal vector so the integration works correctly

source

AdvectionEquationMod.calcBoundaryFunctionalIntegrand — Method.

AdvectionEquationMod.calcBoundaryFunctionalIntegrand

Computes the integrand for boundary functional at a surface SBP node. Every functional needs to have its own method and the functional type determines which method is called.

Inputs

params : eqn.params object

nx : X component of face normal vector

ny : Y component of face normal vector

q : Nodal solution variable

functionalData : Object of the functional being computed

Outputs

functional_integrand : Computed integrand at the surface node

source

PDESolver.evalFunctional — Method.

AdvectionEquationMod.evalFunctional

Hight level function that evaluates functionals specified in the options dictionary. The user must call this function for functional evaluation.This function is agnostic which type of a functional is being computed and calls a mid level type specific function for the actual functional evaluation.

Arguments

mesh : Abstract mesh object

sbp : Summation-By-Parts operator

eqn : Advection equation object

opts : Options dictionary

functionalData : Object of the functional being computed.

source

Functional Types

AdvectionEquationMod.QfluxData — Type.

###AdvectionEquationMod.QfluxData

Data type for storing relevant information pertaining to an a functional or an objective function.

Members

bcnums : boundary condition groups on which the functional is to be computed.

val : Computed value of the functional

target_qFlux : Target value for the functional qFlux

source

AdvectionEquationMod.IntegralQData — Type.

Functional that integrates the solution q over the specified boundary(/ies)

Fields

bcnums: the boundary condition groups the functional is computed over

val: the value of the functional, initially 0.0

source