Common Functions

Advection Common Functions

This page describes functions that are used to calculated quantities that are needed for IC, BCs, and source terms. They are particularly useful for defining manufactures solutions.

AdvectionEquationMod.calc_exp2xplus2y — Method.

AdvectionEquationMod.calc_exp2xplus2y

Calculates and return the expression u = exp(2x + 2y)

source

AdvectionEquationMod.calc_exp3xplusy — Method.

AdvectionEquationMod.calc_exp3xplusy

Calculates and return the expression u = exp(3*x + y)

source

AdvectionEquationMod.calc_exp5xplus4yplus2 — Method.

AdvectionEquationMod.calc_exp5xplus4yplus2

Calculates and returns the expression u = exp(5x + 4y +2)

source

AdvectionEquationMod.calc_exp5xplusy — Method.

AdvectionEquationMod.calc_exp5xplusy

Calculates and return the expression u = exp(5*x + y)

source

AdvectionEquationMod.calc_exp_xplusy — Method.

AdvectionEquationMod.calc_exp_xplusy

Calculates and returns e^(x + y)

source

AdvectionEquationMod.calc_exp_xy — Method.

AdvectionEquationMod.calc_exp_xy

Calculates and returns u = exp(x*y)

source

AdvectionEquationMod.calc_mms1 — Method.

AdvectionEquationMod.calc_mms1

Calculates and returns the value of the solution for doing Method of Manufactured solutions. This is for debugging only, and could change at any time.

source

AdvectionEquationMod.calc_mms1dx — Method.

AdvectionEquationMod.calc_mms1dx

Calculates and returns the x derivative of calc_mms1

source

AdvectionEquationMod.calc_p0 — Method.

calculates and returns a degree 0 polynomial (ie. a constant)

source

AdvectionEquationMod.calc_p1 — Method.

AdvectionEquationMod.calc_p1

Calculates and returns a 1st order polynomial of x and y

source

AdvectionEquationMod.calc_p1dx — Method.

AdvectionEquationMod.calc_p1dx

Calculates and returns the x derivative of calc_p1

source

AdvectionEquationMod.calc_p1dy — Method.

AdvectionEquationMod.calc_p1dy

Calculates and returns the y derivative of calc_p1

source

AdvectionEquationMod.calc_p1dz — Method.

AdvectionEquationMod.calc_p1dz

Calculates and returns the z derivative of calc_p1

source

AdvectionEquationMod.calc_p2 — Method.

AdvectionEquationMod.calc_p2

Calculates and returns a 2nd order polynomial in x and y

source

AdvectionEquationMod.calc_p2dx — Method.

AdvectionEquationMod.calc_p2dx

Calculates and returns a the x derivative of calc_p2

source

AdvectionEquationMod.calc_p2dy — Method.

AdvectionEquationMod.calc_p2dy

Calculates and returns the y derivative of calc_p2

source

AdvectionEquationMod.calc_p2dz — Method.

AdvectionEquationMod.calc_p2dz

Calculates and returns the z derivative of calc_p2

source

AdvectionEquationMod.calc_p3 — Method.

AdvectionEquationMod.calc_p3

Calculates and returns a 3rd order polynomial in x and y (and z in 3d)

source

AdvectionEquationMod.calc_p3dx — Method.

AdvectionEquationMod.calc_p3dx

Calculates and returns the x derivataive of calc_p3

source

AdvectionEquationMod.calc_p3dy — Method.

AdvectionEquationMod.calc_p3dy

Calculates and returns the y derivative of calc_p3

source

AdvectionEquationMod.calc_p3dz — Method.

AdvectionEquationMod.calc_p3dz

Calculates and returns the z derivative of calc_p3

source

AdvectionEquationMod.calc_p4 — Method.

AdvectionEquationMod.calc_p4

Calculates and returns a 4th order polynomial in x and y

source

AdvectionEquationMod.calc_p4dx — Method.

AdvectionEquationMod.calc_p4x

Calculates and returns the x derivative of calc_p4

source

AdvectionEquationMod.calc_p4dy — Method.

AdvectionEquationMod.calc_p4dy

Calculates and returns the y derivative of calc_p4

source

AdvectionEquationMod.calc_p4dz — Method.

AdvectionEquationMod.calc_p4dz

Calculates and returns the z derivative of calc_p4

source

AdvectionEquationMod.calc_p5 — Method.

AdvectionEquationMod.calc_p5

Calculates and returns a 5th order polynomial in x and y (and z in 3d)

source

AdvectionEquationMod.calc_p5dx — Method.

AdvectionEquationMod.calc_p5dx

Calculates and returns the x derivative of calc_p5

source

AdvectionEquationMod.calc_p5dy — Method.

AdvectionEquationMod.calc_p5y

Calculates and returns the y derivative of calc_p5

source

AdvectionEquationMod.calc_p5dz — Method.

AdvectionEquationMod.calc_p5z

Calculates and returns the z derivative of calc_p5

source

AdvectionEquationMod.calc_sinwave — Method.

AdvectionEquationMod.calc_sinwave

Calculates and returns sin(-x + t)

source

AdvectionEquationMod.calc_sinwave_ampl — Method.

AdvectionEquationMod.calc_sinwave_ampl

Calculates and returns Asin(-x + omegat)

source

AdvectionEquationMod.calc_sinwavey — Method.

AdvectionEquationMod.calc_sinwavey

Calculates and returns sin(y)^2 + 5*sin(y) + 3/sin(y)

source

AdvectionEquationMod.calc_sinwavey_pert — Method.

AdvectionEquationMod.calc_sinwavey_pert

Calculates and returns 1000sin(x)calc_sinwavey

source

AdvectionEquationMod.calc_unsteadymms — Method.

u = exp(x + y + z + t) in 3d (z = 0 in 2d)

source

AdvectionEquationMod.calc_x4 — Method.

AdvectionEquationMod.calc_x4

Calculates and returns a 4th order polynomial in x

source

AdvectionEquationMod.calc_x4der — Method.

AdvectionEquationMod.calc_x5plusy5

Calculates and returns the x derivative of calc_x4

source

AdvectionEquationMod.calc_x5plusy5 — Method.

AdvectionEquationMod.calc_x5plusy5

Calculates and returns x^5 + y^5

source

AdvectionEquationMod.calc_xplusy — Method.

AdvectionEquationMod.calc_x5plusy5

Calculates and returns u = x+y

source