Advection Source Term

This pages describes the functions that apply source terms

AdvectionEquationMod.getSRCFunctors — Method.

This function gets the functor specified by opts["SRCname"] and stores it to the equation object. Currently one 1 source functor is allowed.

AdvectionEquationMod.SRCDict — Constant.

AdvectionEquationMod.SRCDict

It stores all the possible boundary condition dictionary options. Whenever a new boundary condition is created, it should get added to BCDict.

All functors must have the signature:

src_func(params, coords::ParamType, t)

where coords is the vector of length 2 containing the x and y coordinates of the node, params.alpha_x and params.alpha_y are the advection velocities in the x an y directions, and t is the current time

AdvectionEquationMod.SRC0 — Type.

AdvectionEquationMod.SRC0

This is the zero source term. This is the default of source term is specified

AdvectionEquationMod.SRC1 — Type.

AdvectionEquationMod.SRC1

This source term returns 1 everywhere.

AdvectionEquationMod.SRC2 — Type.

AdvectionEquationMod.SRC2

This source term returns 1 everywhere.

AdvectionEquationMod.SRCexp2xplus2y — Type.

AdvectionEquationMod.SRCexp2xplus2y

Calculates the source term for q = exp(2x + 2y)

AdvectionEquationMod.SRCexp3xplusy — Type.

AdvectionEquationMod.SRCexp3xplusy

Calculates the source term for q = exp(3*x + y)

AdvectionEquationMod.SRCexp5xplus4yplus2 — Type.

AdvectionEquationMod.SRCexp5xplus4yplus2

Calculates the source term for q = exp(5x + 4y +2)

AdvectionEquationMod.SRCexp5xplusy — Type.

AdvectionEquationMod.SRCexp5xplusy

Calculates the source term for q = exp(5*x + y)

AdvectionEquationMod.SRCexp_xplusy — Type.

AdvectionEquationMod.SRCexp_xplusy

This is a source term that returns a source term for e^(x+y)

AdvectionEquationMod.SRCexp_xy — Type.

AdvectionEquationMod.SRCexp_xy

Calculates the source term for q = exp(x*y)

AdvectionEquationMod.SRCmms1 — Type.

AdvectionEquationMod.SRC1

This source term that returns: the derivative of mms1

AdvectionEquationMod.SRCp1 — Type.

AdvectionEquationMod.SRCp1

This source term that returns: the source term for a manufactured solution using a 1st order polynomial

AdvectionEquationMod.SRCp2 — Type.

AdvectionEquationMod.SRCp3

This source term that returns: the source term for a manufactured solution using a 2nd order polynomial

AdvectionEquationMod.SRCp3 — Type.

AdvectionEquationMod.SRCp3

This source term that returns: the source term for a manufactured solution using a 3rd order polynomial

AdvectionEquationMod.SRCp4 — Type.

AdvectionEquationMod.SRCp4

This source term that returns: the source term for a manufactured solution using a 4th order polynomial

AdvectionEquationMod.SRCp5 — Type.

AdvectionEquationMod.SRCp5

This source term that returns: the source term for a manufactured solution using a 5th order polynomial

AdvectionEquationMod.SRCunsteadymms — Type.

Source term for unsteady mms

AdvectionEquationMod.SRCunsteadypoly — Type.

Source term for unsteady poly

AdvectionEquationMod.SRCx — Type.

AdvectionEquationMod.SRCx

This source term that returns: f = x

AdvectionEquationMod.SRCx4 — Type.

AdvectionEquationMod.SRCx4

This source term that returns: the source term for a manufactured solution using a 4th order polynomial

AdvectionEquationMod.SRCx5plusy5 — Type.

AdvectionEquationMod.SRCx5plusy5

This is a source term that returns a source term for e^(x+y)

AdvectionEquationMod.SRCxplusy — Type.

AdvectionEquationMod.SRCxplusy

calculates the source term for q = x + y