Advection Types

This page provides the documentation for the DataTypes defined in the Advection module

AdvectionEquationMod.ParamType — Type.

Subtype of AbstractParamType.

Static Parameters:

Tsol

Tres

Tdim

This is a container passed to all low level function, useful for storing miscellaneous parameters or constants

AdvectionEquationMod.ParamType2 — Constant.

Convenient alias for all 2D ParamTypes

AdvectionEquationMod.ParamType3 — Constant.

Convenient alias for all 3D ParamTypes

AdvectionEquationMod.AbstractAdvectionData — Type.

Direct subtype of AbstractSolutionData, inheriting Tsol and Tres as static parameter

AdvectionEquationMod.AdvectionData — Type.

Subtype of AbstractAdvectionData, inheriting its static parameters and adding Tdim.

AdvectionEquationMod.AdvectionData_ — Type.

AdvectionEquationMod.AdvectionData_

This type is an implementation of the abstract AdvectionData. It is parameterized by Tsol, the datatype of the solution variables and Tmsh, the datatype of the mesh variables. Tres is the 'maximum' type of Tsol and Tmsh. Tdim is the dimensionality of the equation being solve (2 or 3).

This type is (ultimately) a subtype of AbstractSolutionData and contains all the required fields.

AdvectionEquationMod.PhysicsName — Constant.

This physics is named Advection