Euler Equation Steady Adjoint

PDESolver currently has the capability to compute the steady adjoint of a boundary functional. Recall the adjoint equation as

\[\frac{\partial \mathcal{L}}{\partial q} = \frac{\partial \mathcal{J}}{\partial q} + \psi^T \frac{\partial \mathcal{R}}{\partial q} = 0\]

where, $\mathcal{L}$ is the Lagrangian for functional $\mathcal{J}$ and $q$ is the solution variable. The adjoint can be computed by calling the function calcAdjoint, which has been described below.

EulerEquationMod.calcAdjoint — Method.

EulerEquationMod.calcAdjoint

Calculates the adjoint vector, ψ, for a single functional. The user must always call this function in order to compute the adjoint vector. Currently only DG meshes are supported. The function performs a direct solve using Julia's \ operator. For parallel meshes, a PETSc solve is done using ILU factorization.

Inputs

mesh : Abstract DG mesh type

sbp : Summation-By-parts operator

eqn : Euler equation object

opts : Options dictionary

functionalData : Object corresponding the boundary functional being computed. It must be a subtype of AbstractOptimizationData

adjoint_vec : Resulting adjoint vector. In the parallel case, the adjoint vector is distributed over the processors similar to eqn.q_vec i.e. every rank has its share of the adjoint vector corresponding to the dofs on the rank.

functional_number : Numerical identifier to obtain geometric edges on which a functional acts

Outputs

None

EulerEquationMod.calcFunctionalDeriv — Method.

EulerEquationMod. calcFunctionalDeriv

Computes a 3D array of the derivative of a functional w.r.t eqn.q on all mesh nodes.

Inputs

mesh : Abstract DG mesh type

sbp : Summation-By-parts operator

eqn : Euler equation object

opts : Options dictionary

functionalData : Functional object of super-type AbstractOptimizationData that is needed for computing the adjoint vector. Depending on the functional being computed, a different method based on functional type may be needed to be defined.

func_deriv_arr : 3D array that stores the derivative of the functional w.r.t. eqn.q. The array is the same size as eqn.q

Outputs

None

EulerEquationMod.calcIntegrandDeriv — Method.

EulerEquationMod.calcIntegrandDeriv

Compute the derivative of the functional Integrand at a node w.r.t all the degrees of freedom at the node.

Inputs

opts : Options dictionary

params : parameter type

q : Solution variable at a node

aux_vars : Auxiliary variables

nrm : normal vector in the physical space

integrand_deriv : Derivative of the integrand at that particular node

node_info : Tuple containing information about the node

functionalData : Functional object that is a subtype of AbstractOptimizationData.

Outputs

None

EulerEquationMod.calcResidualJacobian — Method.

###EulerEquationMod.calcResidualJacobian

The function calculates the residual for computing the adjoint vector. The function allows for jacobian to be computed depending on the jacobian type specified in the options dictionary jac_type.

Input

mesh : Abstract mesh object

sbp : Summation-By-parts operator

eqn : Euler equation object

opts : Options dictionary

Output

jac : Jacobian matrix