Startup

This page documents the functions that start the solver.

Functions

EulerEquationMod.createObjects — Method.

This function creates and initializes the mesh, sbp, eqn, and opts objects

Inputs: opts: options dictionary

Outputs: mesh: an AbstractMesh. The concrete type is determined by the options dictionary sbp: an AbstractSBP. The concrete type is determined by the options dictionary eqn: an EulerData object opts: the options dictionary pmesh: mesh used for preconditioning, can be same object as mesh

EulerEquationMod.createObjects — Method.

Constructs a the EulerData object given a mesh, sbp, and options dictionary.

Used for submesh solves.

Inputs

mesh

sbp

opts

Outputs

mesh

sbp

eqn

opts

pmesh: currently, always the same as mesh

EulerEquationMod.postproc — Method.

This function does post processing, if requested by the input options. Typical post processing includes calculation of errors, norms of important quantities, writing of files. etc.

Inputs: mesh sbp eqn opts

PDESolver.solvePDE — Function.

Given fully initialized mesh, sbp, eqn, opts, this function solves the Euler equations. The 4 object should be obtained from createObjects().

Specifically, it applies an initial condition and invokes a nonlinear solver according to the options dictionary.

Inputs: mesh: an AbstractMesh sbp: an AbstractSBP eqn: an AbstractEulerData opts: the options dictionary. This must be the options dictionary returned by createObjects(). Changing values in the options dictionary after calling createObjects() results in undefined behavior. pmesh: mesh used for preconditioning, can be same object as mesh. default value of mesh

EulerEquationMod.checkOptions — Method.

This function checks the options in the options dictionary after default values have been supplied and throws exceptions if unsupported/incompatable options are specified

Inputs: opts: options dictionary

Outputs: none