Input

A PDESolver execution is controlled by an options dictonary. The user must supply this dictonary in the form of a Julia source file that declares a Dict{Any, Any} called arg_dict. The file path (relative to the users pwd must be passed to the solver as the the first optional argument. For example, to launch a PDESolver run, execute

julia /path/to/startup.jl "path/to/dictonary"

Descriptions of the valid keys and values can be found in the file input_vals.txt. PDESolver supplies default values for all keys for which there is a sane default. Values that might be unused are initialized to -1 or "none".

Conventions

Physics modules generally use the majorIterationCallback function to log important quantities to files. Such logging should be controlled by two keys, "write_outname" where outname is the name of the quantity, which has a boolean value, and "write_outname_fname" that has a string value containing the name of the file to write (including extension). Examples out things that can be logged are entropy and kinetic energy. Both these keys should have default values, and users should generally not need to modify the second one.

Key Validation

After supplying default values, PDESolver checks that all keys in the dictonary are recognized keys. It does this by comparing against the list of keys documented in the input_vals.txt file. A warning of is printed to STDERR if an unrecognized key is found.

The mechanics of the key validation are as follows. The extract_keys.jl script reads the input_vals.txt file (and the input_vals_internal.txt), looking for any string that is surrounded by double quotes and starts in the first character of a line. All keys are written to a dictonary in the file known_keys.jl. This file is included by PDESolver.

The script extract_keys.jl is run as part of PDESolver installation. If new keys are added it must be run again to silence warnings during key validation.

Input.make_input — Method.

Make an input file from an options dictionary.

Inputs: dict: the options dictionary fname: the file name (without extension)

Input.read_input — Method.

PDESolver.read_input

This function reads a file which must be a julia source file that declares a dictionary of option keywords and values for the options named arg_dict. See the documention on input variables for valid keywords.

read_input() returns the dictionary after doing some sanity checks and supplying default values for any unspecified keys.

After supplying default values, it prints the dictonary to arg_dict_output.jl, which is a valid julia source file and can be read in to re-run a simulation.

This function checks whether the keys in arg_dict are recognized keywords and prints a warning to STDERR if an unrecognized key is found. The list of known keys is read from the julia source file known_keys.jl

Inputs: * fname : name of file to read

Outputs: arg_dict: a Dict{Any, Any} containing the option keywords and values

Input.checkForIllegalOptions_post — Method.

Check the user supplied options for errors after supplying default options.

Input.checkForIllegalOptions_pre — Method.

Check the user supplied options for errors before supplying default values

Input.checkKeys — Method.

PDESolver.checkKeys

This function verifies all the keys in the first argument are also keys of the second argument and prints a warning to STDERR if they are not.

Inputs arg_dict: first dictonary known_keys: second dictonary

Outputs: cnt: number of unrecognized keys