Documentation of the PDESolver Physics Module Interface

The PDESolver Module provides some structure for physics modules to plug into. See Interfaces in PDESolver for details. Each physics module should extend evalResidual with a new method to which takes the AbstractSolutionData defined by the physics module as an argument. This structure allows evalResidual to be visible to the NonlinearSolver module while being defined by the physics modules.

Evaluating the Physics

PDESolver.evalResidual — Function.

This function evalutes dq/dt = R(q). For steady problems it evalutes R(q) at some state q. The state is stored in eqn.q, and eqn.res is populated with R(q). Note that these are 3 dimensional arrays. The physics modules only interact with the 3 dimensional arrays, never the vectors eqn.q_vec and eqn.res_vec. Each physics module must implement this function for its own subtype of AbstractSolutionData (ie. with a more specific type for the eqn argument and equallty specific types for the other arguments). This is important because evalResidual is common to all physics modules, so a user or some other part of the code can call evalResidual(mesh, sbp eqn, opts), and Julia's multiple dispatch will figure out the right method to call based on the type of the eqn argument.

The evaluation of the residual R(q) should depend only on the data stored in mesh, sbp, eqn, and opts, and any data that depend on q should be recalculated every time the function is called. This function is used as a building block by other parts of the solver, particularly the NonlinearSolvers. See interfaces.md for details

Inputs: mesh: an AbstractMesh describing the mesh on which to solve the physics sbp: an SBP operator eqn: a subtype of AbstractSolution data, used to store all of the data used by the physics module opts: the options dictionary t: the current time value, defaults to 0.0

Registration Functions

These function provide the basic API for registering and retrieving physics modules with the PDESolver module.

PDESolver.register_physics — Function.

This function registered a new physics module with the global list of all known physics modules. Every physics module should do this as part of module initialization. The name, the module, and the startup function must be unique (ie. they must not already exist in the list). This function throws and exception if they are not.

Inputs:

modname:  an ASCIIString name for this entry in the list.  It is used
          to retrieve the module and startup function in the 
          retrieve_physics function. Typically the name is capitalized.
mod:  the Module itself
startup_func: the function for running the physics.  It must have signature
              startup_func(fname::ASCIIString), where fname is the name of
              an input file

Outputs:

none

PDESolver.retrieve_physics — Function.

Retrieves the physics module and function registered using register_physics

Input:

modname: an ASCIIString containing the name of the module supplied to
         `register_physics`

Outputs:

mod: the physics Module
func: the function to evaluate the physics

See also registerIC and registerBC.