Euler Types

This page provides documentations for DataTypes and and simple functions that are defined in the Euler module

EulerEquationMod.EulerData_ — Type.

This type is an implimentation of the abstract EulerData. It is paramterized by the residual datatype Tres and the mesh datatype Tmsh because it stores some arrays of those types. Tres is the 'maximum' type of Tsol and Tmsh, where Tsol is the type of the conservative variables. It is also paremterized by var_type, which should be a symbol describing the set of variables stored in eqn.q. Currently supported values are :conservative and :entropy, which indicate the conservative variables and the entropy variables described in:

'A New Finite Element Formulation for Computational Fluid Dynamics: Part I' by Hughes et al.`

Eventually there will be additional implimentations of EulerData, specifically a 3D one.

Static Parameters:

Tsol : datatype of variables solution variables, ie. the q vector and array

Tres : datatype of residual. ie. eltype(res_vec)

Tdim : dimensionality of equation, integer, (2 or 3, currently only 2 is supported).

Tmsh : datatype of mesh related quantities

var_type : symbol describing variables used in weak form, (:conservative or :entropy)

EulerEquationMod.AbstractEulerData — Type.

EulerEquationMod.AbstractEulerData{Tsol, Tres}

This abstract type should be the supertype of all solution data objects that are related to the Euler equations.

It should be used for specify the type of a function argument only when the function does no operations on the solution data object itself, it just passes it onto other functions that do the work (thus AbstractEulerData should be used for only the highest level functions).

Another way of saying say it is that this type should only be used when the function only needs to ensure that it is solving the Euler equations, but does not care even a little bit about how.

EulerEquationMod.EulerData — Type.

EulerEquationMod.EulerData

This type, although abstract, is the type functions should use for their input arguments if they do any operations on the solution data object. It stores all data used in evaluting the Euler Equations.

It is paramaterized on the types Tsol, the type of the conservative variables q, and Tdim, the dimension of the equation

It should have the following fields: * res_type : datatype of residual (depreciated) * q : 3D array holding conservative variables * q_vec : vector to assemble q into * aux_vars : 3D array holding auxiliary variables * flux_parametric : 4D array [ndof per node, nnodes per element, nelements, Tdim] holding the Euler flux in the xi and eta directions * res : 3D array holding residual * res_vec : vector form of res * edgestab_alpha : paramater used for edge stabilization, 4d array * bndryflux : 3D array holding boundary flux data * stabscale : 2D array holding edge stabilization scale factor * M : vector holding the mass matrix * Minv : vector holding inverse mass matrix # Minv3D : 3D array holding inverse mass matrix for application to res (not res_vec)

EulerEquationMod.BoundaryForceData — Type.

###EulerEquationMod.BoundaryForceData

Composite data type for storing data pertaining to the boundaryForce. It holds lift and drag values

EulerEquationMod.ParamType — Type.

This type holds the values of any constants or paramters needed during the computation. These paramters can be specified in the opts dictionary or have default values set here. If there is no reasonable default, values are initialized to -1

There are also a bunch of arrays that are used as temporaries by low level functions (to avoid having to allocate arrays themselves, which is a performance trap). In general, this Type is used as a container to pass around values.

gamma and R are the independent themodynamic variables

Whether this type should be immutable or not is an open question

This type is paramaterized on the dimension of the equation for purposes of multiple dispatch

Static Parameters:

Tdim : dimensionality of the equation, integer, (used for dispatch)

var_type : type of variables used used in the weak form, symbol, (used for dispatch), currently supported values: :conservative, :entropy

Tsol : datatype of solution variables q

Tres : datatype of residual

Tmsh : datatype of mesh related quantities (mapping jacobian etc.)

Fields (with default values):

cv : specific heat constant

R : specific gas constant (J/(Kg*K))

gamma : ratio of specific heats

gamma_1 : gamma - 1

Fields (without default values):

Ma : free stream Mach number

Re : free stream Reynolds number

aoa : angle of attack (radians)

EulerEquationMod.ParamType2 — Type.

Useful alias for 2D ParamType

EulerEquationMod.ParamType3 — Type.

Useful alias for 3D ParamType

EulerEquationMod.cleanup — Method.

This function performs all cleanup activities before the run_physics() function returns. The mesh, sbp, eqn, opts are returned by run_physics() so there is not much cleanup that needs to be done, mostly closing files.

Inputs/Outputs:

* mesh: an AbstractMesh object
* sbp: an SBP operator
* eqn: the EulerData object
* opts: the options dictionary

EulerEquationMod.openLoggingFiles — Method.

This function opens all used for logging data. In particular, every data file that has data appended to it in majorIterationCallback should be opened here. Most files are of type BufferedIO, so they must be flushed periodically.

This function requires each output to have two keys: "write_outname" and "write_outname_fname", where the first has a boolean value that controls whether or not to write the output, and the second is the file name (including extension) to write.

This function contains a list of all possible log files. Every new log file must be added to the list

Inputs:

mesh: an AbstractMesh (needed for MPI Communicator)

opts: options dictionary

Outputs:

* file_dict: dictionary mapping names of files to the file object
             ie. opts["write_entropy_fname"] => f

Exceptions: this function will throw an exception if any two file names are the same

EulerEquationMod.DragData — Type.

EulerEquationMod.DragData

Subtype of AbstractOptimizationData. Stores all the information relevant to computing an objective function pertaining to drag. Presently its an empty type

EulerEquationMod.FaceElementIntegralType — Type.

Functor type for faceElementIntegrals. These integrals operate on a face, but require data from the entirety of the elements that make up the face, rather than data interpolated to the face

EulerEquationMod.LiftData — Type.

EulerEquationMod.LiftData

Subtype of AbstractOptimizationData. Stores all the information relevant to computing an objective function pertaining to lift. Presently its an empty type

EulerEquationMod.PressureData — Type.

EulerEquationMod.PressureData

Subtype of AbstractOptimizationData. Stores all the information relevant to computing an objective function pertaining to pressure coefficeint

Members

targetCp_arr : An array of arrays that stores the target coefficient of pressure. length(targetCp_arr) = number of geometric edges over which the functional is being computed. Each sub array has dimensions (sbpface.numnodes, nfaces) (from calcBoundarFlux in bc.jl)

nodal_info : 1D array of indices for one node needed to acces targetCp_arr at a particular data point. nodal_info[1] = geometric edge number nodal_info[2] = sbpface node number nodal_info[3] = element face number on the geometric edge

EulerEquationMod.createFunctionalData — Function.

###EulerEquationMod.createFunctionalData

Creates an object for functional computation. This function needs to be called the same number of times as the number of functionals EXCLUDING the objective function are being computed

Arguments

mesh : Abstract PUMI mesh

sbp : Summation-by-parts operator

eqn : Euler equation object

opts : Options dictionary

functional_number : Which functional object is being generated. Default = 1

EulerEquationMod.createObjectiveFunctionalData — Method.

###EulerEquationMod.createObjectiveFunctionalData

Function for create an object for functional and adjoint computation where the functional is an objective function in an optimization.

Arguments

mesh : Abstract PUMI mesh

sbp : Summation-by-parts operator

eqn : Euler equation object

opts : Options dictionary